Addressed some issues detected by users feedback
+Added subtitles in the phone voice message
+Added controls menu
+Changed collider of the blood stain
+Improved water reflection
+Keys are now more visible
+Cursor does not show anymore in object viewer UI
+Fixed typo in the final newspaper document
+Added disclaimer at the start of the game

+++ Assets/Scripts/UI/DialogManagerUI.cs
@@ -37,7 +37,7 @@ public class DialogManagerUI : MonoBehaviour
         canvasUI.SetActive(false);
     }
 
-    public void showText(string text, int secondsToFade)
+    public void showText(string text, float secondsToFade)
     {
         timeToFade = secondsToFade;
         realTime = 0;
